% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_centroid_data}
\alias{get_centroid_data}
\title{Compute the centroids for a given clustering}
\usage{
get_centroid_data(x, k, use_median = FALSE)
}
\arguments{
\item{x}{object obtained from \code{\link{get_kmeans}}}

\item{k}{number of clusters to use}

\item{use_median}{should median be used instead of mean to compute the centroid?}
}
\description{
Compute the centroids for a given clustering
}
