% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{get_kmeans}
\alias{get_kmeans}
\title{Get a set k-means++ clusters for a collection of series}
\usage{
get_kmeans(dat, x, y, groups = NULL, nk = 20)
}
\arguments{
\item{dat}{data frame}

\item{x}{string name of x variable}

\item{y}{string name of y variable}

\item{groups}{a character vector of grouping variable names}

\item{nk}{number of clusters to run through (result be a set of clusters wit k = 2:nk)}
}
\description{
Get a set k-means++ clusters for a collection of series
}
