% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesclust-package.R
\docType{data}
\name{nasd16}
\alias{nasd16}
\title{"Nasdaq 2016" dataset}
\usage{
nasd16
}
\description{
2016 monthly median closing price for 2801 stocks in the Nasdaq exchange.
}
\examples{
\dontrun{
library(dplyr)

# scale the monthly median close price so that we are clustering on general shape
d <- nasd16 \%>\%
  group_by(symbol) \%>\%
  mutate(close_scl = scale(med_close)) \%>\%
  select(-company, -med_close)

set.seed(1234)
# k-means clustering with 2, 5, and 9 clusters
km <- get_kmeans(d, x = "month", y = "close_scl", k = c(2, 5, 9))
plot_scree(km)
plot_heat(km, 9, col = "sector")
plot_heat(km, 9, col = "sector", interactive = FALSE)
plot_heat(km, 9, col = "industry", cutoff = 10)
}
}
\keyword{data}
