% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{plot_clust}
\alias{plot_clust}
\title{Plot clustered series}
\usage{
plot_clust(x, k, centroid = TRUE, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x}{object obtained from \code{\link{get_kmeans}}}

\item{k}{number of clusters to plot}

\item{centroid}{should centroids be overlaid?}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{\ldots}{additional parameters passed on to \code{\link{xyplot}}}
}
\description{
Plot clustered series
}
