% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{plot_heat}
\alias{plot_heat}
\title{Plot heat map of variable distributions within each cluster}
\usage{
plot_heat(x, k, col, cutoff = 1, color = viridis::magma,
  xaxis_height = NULL, yaxis_width = NULL, interactive = TRUE, ...)
}
\arguments{
\item{x}{object obtained from \code{\link{get_kmeans}}}

\item{k}{number of clusters to use}

\item{col}{name of variable to plot (either categorical or sliced numeric variable in the data)}

\item{cutoff}{minimum number of series that must be present for a level of the variable to be included in the plot}

\item{color}{a vector of colors or a color ramp function}

\item{xaxis_height, yaxis_width}{size of axes, in pixels (ignored if \code{interactive = FALSE})}

\item{interactive}{if \code{TRUE}, \code{\link[d3heatmap]{d3heatmap}} will be used, else \code{\link[pheatmap]{pheatmap}}}

\item{\ldots}{parameters passed on to \code{\link[d3heatmap]{d3heatmap}} or \code{\link[pheatmap]{pheatmap}}}
}
\description{
Plot heat map of variable distributions within each cluster
}
\examples{
\dontrun{
library(dplyr)

# scale the monthly median close price so that we are clustering on general shape
d <- nasd16 \%>\%
  group_by(symbol) \%>\%
  mutate(close_scl = scale(med_close)) \%>\%
  select(-company, -med_close)

set.seed(1234)
# k-means clustering with 2, 5, and 9 clusters
km <- get_kmeans(d, x = "month", y = "close_scl", k = c(2, 5, 9))
plot_scree(km)
plot_clust(km, 9)
plot_heat(km, 9, col = "sector")
plot_heat(km, 9, col = "sector", interactive = FALSE)
plot_heat(km, 9, col = "industry", cutoff = 10)
}
}
