% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{plot_scree}
\alias{plot_scree}
\title{Scree plot}
\usage{
plot_scree(x, log = FALSE)
}
\arguments{
\item{x}{object obtained from \code{\link{get_kmeans}}}

\item{log}{plot the y-axis on the log scale?}
}
\description{
Scree plot
}
