% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{print.pheatmap_plot}
\alias{print.pheatmap_plot}
\title{Print trellis plot to servr}
\usage{
\method{print}{pheatmap_plot}(x, ...)
}
\arguments{
\item{x}{pheatmap_plot}

\item{\ldots}{additional parameters}
}
\description{
Print trellis plot to servr
}
