% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{plot_clust}
\alias{plot_clust}
\title{Plot clustered series}
\usage{
plot_clust(x, k, centroid = TRUE, xlab = NULL, ylab = NULL, alpha = 0.3,
  use_median = FALSE, ...)
}
\arguments{
\item{x}{object obtained from \code{\link{get_kmeans}}}

\item{k}{number of clusters to plot}

\item{centroid}{should centroids be overlaid?}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{alpha}{alpha to use for lines}

\item{use_median}{should median be used instead of mean to compute the centroid?}

\item{\ldots}{additional parameters passed on to \code{\link{xyplot}}}
}
\description{
Plot clustered series
}
\examples{
\dontrun{
library(dplyr)

# scale the monthly median close price so that we are clustering on general shape
d <- nasd16 \%>\%
  group_by(symbol) \%>\%
  mutate(close_scl = as.numeric(scale(med_close))) \%>\%
  select(-company, -med_close) \%>\%
  ungroup()

set.seed(1234)
# k-means clustering with 2, 5, and 9 clusters
km <- get_kmeans(d, x = "month", y = "close_scl", k = c(2, 5, 9, 25))
plot_scree(km)
plot_heat(km, 9, col = "sector")
plot_heat(km, 9, col = "sector", interactive = FALSE,
  display_numbers = TRUE, cutree_cols = 3, cutree_rows = 3)
plot_heat(km, 9, col = "industry", cutoff = 20)

heat <- plot_heat(km, 9, col = "sector", interactive = FALSE,
  display_numbers = TRUE, cutree_cols = 3, cutree_rows = 3,
  annotation_labs = c("A", "B", "C"))
heat

cents <- get_centroid_data(km, 9)
plot_centroid_groups(cents, heat)
}
}
