% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat.R
\name{plot_heat}
\alias{plot_heat}
\title{Plot heat map of variable distributions within each cluster}
\usage{
plot_heat(x, col, cutoff = 1, cluster_rows = TRUE, cluster_cols = FALSE)
}
\arguments{
\item{x}{object obtained from \code{\link{get_kmeans}}}

\item{col}{name of variable to plot (either categorical or sliced numeric variable in the data)}

\item{cutoff}{minimum number of series that must be present for a level of the variable to be included in the plot}

\item{cluster_rows}{should the rows be clustered in the \code{\link{pheatmap}} plot?}

\item{cluster_cols}{should the columns be clustered in the \code{\link{pheatmap}} plot?}
}
\description{
Plot heat map of variable distributions within each cluster
}
